# docs/swagger.yaml
definitions:
...
  user.UpdatePayKeyRequest:
    properties:
      pay_key:
        maxLength: 10
        type: string
    required:
    - pay_key
    type: object
...
  util.ResponseAny:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
info:
  contact: {}
  title: LINUX DO PAY
  version: 0.1.0
paths:
...
  /api/v1/user/pay-key:
    put:
      consumes:
      - application/json
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdatePayKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - user
swagger: "2.0"